GREY-40: Correct autodetection of database
This is a minimalistic approach to correcting the issue that without a
database name on the URL default DB cannot be relied on existing.

A better solution than attempting to create the DB and failing on line
145 would be to execute 
"SELECT 1 FROM pg_database WHERE datname = 'indaba';"
and if it returns a value the DB exists.  then creation can be attempted
if appropriate.

I would note that the bootstrap process is dependent on the user having
access to the appropriate tables and databases.  In order to account for
the situation where the user performing the connection does not have any
access except to the specified database a connection should be attempted
to there first and if it fails then attempt connecting to postgres.