Update to constants configuration
Update constants configuration to reflect the updated set of constants
when building a docker image.

Change greyscale.rest.js to not include port if the port is not
configured.  This allows for natural URLS
(http://amida-tech.com/something) without explicitly listing port when
not needed.  This can be important for corst since http://f.com/ is
different from http://f.com:80/